+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Starlight charger displaying incorrect/negative stored charge on servers
+
 ## [2.6.17] - 2021-04-04
 ### Fixed
 - Starlight charger applying multiple instances of enchantment at different levels [#360]