+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.6.18] - 2021-04-07
 ### Changed
 - Reworded the "incorrect network version" message to hopefully be clearer [#367]
 ### Fixed